requirements.txt
absl-py==2.2.2
addict==2.4.0
affine==2.4.0
aioboto3==14.1.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
bokeh==2.4.3
boto3==1.37.1
botocore==1.37.1
cachetools==5.5.2
casadi==3.7.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cmake==4.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
control==0.9.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
descartes==1.1.0
diffusers==0.33.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.14.0
fiona==1.10.1
fire==0.7.0
flash-attn==2.6.3
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.2.0
grpcio==1.71.0
guppy3==3.1.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.2.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.0
jupyterlab-pygments==0.3.0
kiwisolver==1.4.7
lightning-utilities==0.14.3
lit==18.1.8
mmcv-full==1.7.1
markdown==3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmcls==0.25.0
mmdet==2.28.2
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.4.0
notebook-shim==0.2.4
nuscenes-devkit==1.1.11
omegaconf==2.3.0
opencv-python==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
positional-encodings==6.0.1
prettytable==3.16.0
prometheus-client==0.21.1
prompt-toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py==1.11.0
pyarrow==19.0.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.22.0
pygments==2.19.1
pyinstrument==5.0.1
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.6.1
pyquaternion==0.9.9
pysocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==2.2.1
pytz==2023.4
pyyaml==6.0.2
pyzmq==26.4.0
rasterio==1.3.11
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests>=2.28.2
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.24.0
rtree==1.4.0
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.13.1
selenium==4.31.0
send2trash==1.8.3
setuptools==60.2.0
shapely==1.8.5.post1
six==1.17.0
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.6
sqlalchemy==1.4.27
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.6.0
timm==1.0.15
tinycss2==1.4.0
tomli==2.2.1
torchaudio==2.0.2
torchmetrics==1.7.1
torchvision==0.15.2
tornado==6.4.2
tqdm==4.65.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
triton==2.0.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yapf==0.31.0
yarl==1.20.0
zipp==3.21.0
jupyterlab-server
geopandas
# torch==2.0.1+cu118
# numpy==1.23.4
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91

# navsim==1.1.0
# nuplan-devkit==1.2.0

# geopandas==1.0.1